# docs/openapi.yaml
openapi: 3.1.0
info:
  title: piper API
  version: 0.1.0
  description: |
    Project-owned resources are always exposed below `/api/projects/{project_id}`.
    Collection POST operations return `201 Created`, successful mutations with no
    response body return `204 No Content`, JSON field names use snake_case, and
    errors use the shape `{"error":"message"}`.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/projects/{project_id}/schedules:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Schedules.
          headers:
            X-Total-Count: { $ref: "#/components/headers/TotalCount" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Schedule" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, yaml, type]
              properties:
                name: { type: string }
                yaml: { type: string }
                type: { type: string, enum: [immediate, once, cron] }
                cron: { type: string }
                run_at: { type: string, format: date-time }
                params: { type: object, additionalProperties: true }
                max_runs: { type: integer, minimum: 0 }
      responses:
        "201":
          description: Created schedule id.
          content:
            application/json:
              schema:
                type: object
                required: [schedule_id]
                properties:
                  schedule_id: { type: string }
  /api/projects/{project_id}/schedules/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/ScheduleID"
      responses:
        "200":
          description: Schedule detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
    patch:
      parameters:
        - $ref: "#/components/parameters/ScheduleID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: Enabled flag updated.
    delete:
      parameters:
        - $ref: "#/components/parameters/ScheduleID"
      responses:
        "204":
          description: Deleted.
...
  /api/projects:
    get:
      description: Projects visible to the caller (all projects for a system admin).
      responses:
        "200":
          description: Projects.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      description: System admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name]
              properties:
                id: { type: string, pattern: "^[a-z0-9-]+$" }
                name: { type: string }
                description: { type: string }
                owner_member_id: { type: string }
      responses:
        "201":
          description: Created project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /api/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      responses:
        "200":
          description: Project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      description: System admin only. Stops the project's live runs, notebooks, and services before deleting it.
      responses:
        "204":
          description: Deleted.
        "409":
          description: The last remaining project cannot be deleted.
  /api/projects/{project_id}/members:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Project members.
          headers:
            X-Total-Count: { $ref: "#/components/headers/TotalCount" }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectMember" }
    post:
      description: Project admin only. Identify the user by `username` (or `user_id`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                user_id: { type: string }
                username: { type: string }
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "201":
          description: Added member.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMember" }
  /api/projects/{project_id}/members/candidates:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      description: Project admin only. Users who are not yet members.
      responses:
        "200":
          description: Candidate usernames.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username: { type: string }
  /api/projects/{project_id}/members/{user_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
      - name: user_id
        in: path
        required: true
        schema: { type: string }
    patch:
      description: Project admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/ProjectRole" }
      responses:
        "200":
          description: Updated member.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMember" }
    delete:
      description: Project admin only.
      responses:
        "204":
          description: Removed.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: string }
...
    ScheduleID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      description: Page size. When set, the response carries `X-Total-Count`.
      schema: { type: integer, minimum: 1 }
    Offset:
      name: offset
      in: query
      description: Rows to skip; only applies together with `limit`.
      schema: { type: integer, minimum: 0 }
  headers:
    TotalCount:
      description: Total rows matching the filter, ignoring limit/offset. Sent whenever `limit` is.
      schema: { type: integer }
  schemas:
...
    ProjectRole:
      type: string
      enum: [viewer, member, admin]
    ProjectMember:
      type: object
      properties:
        project_id: { type: string }
        user_id: { type: string }
        username: { type: string }
        role: { $ref: "#/components/schemas/ProjectRole" }
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        owner_member_id: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Schedule:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        pipeline_yaml: { type: string }
        template_version_id: { type: string }
        schedule_type: { type: string, enum: [immediate, once, cron] }
        cron_expr: { type: string }
        enabled: { type: boolean }
        max_runs: { type: integer }
        params_json: { type: string }
        last_run_at: { type: string, format: date-time }
        next_run_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...